Project-social-media/app/api/requirements.txt
fastapi==0.111.1
pydantic==2.8.2
torch==2.3.1
diffusers==0.29.2
python-dotenv==1.0.0
langchain==0.2.6
langchain_groq==0.1.6
transformers==4.41.2
accelerate==0.31.0